+++ src/dialect/druidDialect.ts
@@ -15,8 +15,14 @@
  */
 
 import { Duration, Timezone } from 'chronoshift';
+import { NamedArray } from 'immutable-class';
+import { Attributes } from '../datatypes';
 import { PlyType } from '../types';
 import { SQLDialect } from './baseDialect';
+
+export interface DruidDialectOptions {
+  attributes?: Attributes;
+}
 
 export class DruidDialect extends SQLDialect {
   static TIME_PART_TO_FUNCTION: Record<string, string> = {
@@ -63,15 +69,18 @@ export class DruidDialect extends SQLDialect {
     },
     NUMBER: {
       TIME: 'CAST($$ AS BIGINT)',
-      STRING: 'CAST($$ AS FLOAT)',
+      STRING: 'CAST($$ AS DOUBLE)',
     },
     STRING: {
       NUMBER: 'CAST($$ AS VARCHAR)',
     },
   };
 
-  constructor() {
+  private attributes?: Attributes;
+
+  constructor(options: DruidDialectOptions = {}) {
     super();
+    this.attributes = options.attributes;
   }
 
   public dateToSQLDateString(date: Date): string {
@@ -105,6 +114,24 @@ export class DruidDialect extends SQLDialect {
 
   public substrExpression(a: string, position: number, length: number): string {
     return `SUBSTRING(${a},${position + 1},${length})`;
+  }
+
+  public countDistinctExpression(a: string, parameterAttributeName: string | undefined): string {
+    const attribute = NamedArray.findByName(this.attributes || [], parameterAttributeName);
+    const nativeType = attribute ? attribute.nativeType : undefined;
+    switch (nativeType) {
+      case 'HLLSketch':
+        return `APPROX_COUNT_DISTINCT_DS_HLL(${a})`;
+
+      case 'thetaSketch':
+        return `APPROX_COUNT_DISTINCT_DS_THETA(${a})`;
+
+      case 'hyperUnique':
+        return `APPROX_COUNT_DISTINCT(${a})`;
+
+      default:
+        return `COUNT(DISTINCT ${a})`;
+    }
   }
 
   public isNotDistinctFromExpression(a: string, b: string): string {
@@ -166,8 +193,16 @@ export class DruidDialect extends SQLDialect {
     return `POSITION(${substr} IN ${str}) - 1`;
   }
 
-  public quantileExpression(str: string, quantile: string): string {
-    return `APPROX_QUANTILE(${str}, ${quantile})`;
+  public quantileExpression(str: string, quantile: number, parameterAttributeName: string | undefined): string {
+    const attribute = NamedArray.findByName(this.attributes || [], parameterAttributeName);
+    const nativeType = attribute ? attribute.nativeType : undefined;
+    switch (nativeType) {
+      case 'quantilesDoublesSketch':
+        return `APPROX_QUANTILE_DS(${str}, ${quantile})`;
+
+      default:
+        return `APPROX_QUANTILE(${str}, ${quantile})`;
+    }
   }
 
   public logExpression(base: string, operand: string): string {